+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/data/lookup/LookupTableBasedColumnTest.java
@@ -25,7 +25,7 @@ import static net.splitcells.dem.testing.Assertions.requireThrow;
 import static net.splitcells.gel.data.lookup.LookupTables.lookupTable;
 import static net.splitcells.gel.data.table.Tables.table;
 
-public class LookupBasedColumnTest {
+public class LookupTableBasedColumnTest {
     @UnitTest
     public void testSteam() {
         final var attribute = AttributeI.attribute(Integer.class, "attribute");